add the wrap from the graph to the resume components
