+++ js/PostDetails.jsx
@@ -14,7 +14,7 @@ class PostDetails extends Component {
   // production.mqpdw8dnfc.us-east-1.elasticbeanstalk.com
 
   getPostData = () => {
-    const url = `http://production.mqpdw8dnfc.us-east-1.elasticbeanstalk.com:3000/posts/${this.props.id}`;
+    const url = `http://localhost:3000/posts/${this.props.id}`;
     fetch(url)
       .then(response => response.json())
       .then(json => this.setState({ apiData: json }));
